apple_os_installers.py: urllib3 2.1.0 compatibility
Before urllib3 2.1.0, Content-Encoding: x-gzip was not automatically
decoded. urllib3/urllib3 issue 3174, pull request 3176, commit
5fc48e711b33. There were API changes between urllib3 1.x and 2.0 which
may cause some installations to remain on the older version, so this
code may conceivably run with urllib3 1.x, which does not handle x-gzip,
and urllib3 2.1 or later, which do. Since urllib3’s behavior is unknown,
sniff out the gzip magic and assume that it was not decoded if
Content-Encoding: x-gzip is present. This is not a general solution
(it’s not correct for gzipped content served with another layer of gzip
via Content-Encoding), but it’ll do in this case, where the plist
content will never validly begin with the gzip magic number.

Fixes this error:

gzip.BadGzipFile: Not a gzipped file (b'<?')