apple_compression: initial release
This is a wrapper around Apple’s libcompression, via Python ctypes. It
serves both as a command-line program that can compress and decompress
using any algorithm supported by libcompression, and a Python library to
enable interoperability with Apple-specific formats such as LZFSE for
other Python code.

libcompression supports LZ4, zlib, LZMA, LZ4_RAW, and Brotli, as well as
the Apple-specific LZFSE and LZBITMAP.