apple_compression: make interface more like other compression modules
This interface is very similar to the standard Python library’s zlib,
bz2, lzma, and compression.zstd modules.

For the most part, the example of more modern modules such as bz2 and
lzma are followed, except a Decompressor.flush method is provided as in
zlib (it’s absent from more modern modules). This is important to be
able to detect corrputed or truncated streams during decompression,
avoiding an infinte loop, as noted in <compression.h>’s comment on
compression_stream_process.

It is not possible to implement Decompressor.unused_data, because
libcompression does not indicate where in the input data it encountered
and end-of-stream marker.

The custom singleton (_Compression.get) is also removed in favor of
initializing a module-private global _compression variable at load time.
This more traditional approach was chosen because it allows failures
(such as libcompression not being available) to occur by exception at
module import time, which is more predictable.