apple_compression: fix size-limited Decompressor.decompress
This adds a variety of tests covering several features, including:
 - Decompressor.decompress’ max_length parameter,
 - compress and decompress’ max_size parameters,
 - compress and decompress’ scratch_buffer parameters,
 - compress_scratch_buffer_size and decompress_scratch_buffer_size, and
 - bytearray and memoryview input types.

The first of these revealed a bug in this code, which is now fixed. The
second of these revealed bugs and unexpected behaviors in the underlying
library, which are noted in comments in the test.

This also allows the Compressor and Decompressor stream interface to
accept bytearray and memoryview parameters per declared type hints. This
was already permitted by the internal implementation, and brings this
aspect to parity with the buffer interface.